# docs/swagger.yaml
basePath: /
definitions:
  api.MerchantCreateReq:
    description: Request to create a new merchant
    properties:
      merchant_wallet_address:
        type: string
      name:
        type: string
    type: object
  api.MerchantCreateResp:
    description: Response after creating a merchant
    properties:
      api_key:
        type: string
      id:
        type: string
      merchant_wallet_address:
        type: string
    type: object
  api.orderCreateReq:
    properties:
      amount_minor:
        type: integer
      asset:
        description: e.g., "USDC"
        type: string
      chain:
        description: e.g., "polygon-amoy"
        type: string
      idempotency_key:
        type: string
      merchant_id:
        type: string
    type: object
  api.orderCreateResp:
    properties:
      deposit_address:
        type: string
      order_id:
        type: string
      status:
        type: string
    type: object
  api.orderGetResp:
    properties:
      amount_minor:
        type: integer
      asset:
        type: string
      chain:
        type: string
      confirmed_block:
        type: integer
      created_at:
        type: string
      deposit_address:
        type: string
      id:
        type: string
      merchant_id:
        type: string
      paid_at:
        type: string
      status:
        type: string
      tx_hash:
        type: string
    type: object
  api.paymentDetectedReq:
    properties:
      amount_minor:
        description: optional override; if nil, use order.amount_minor
        type: integer
      order_id:
        type: string
      tx_hash:
        type: string
    type: object
  api.paymentDetectedResp:
    properties:
      message:
        type: string
      order_id:
        type: string
      status:
        type: string
    type: object
  api.refundReq:
    properties:
      amount_minor:
        type: integer
      order_id:
        type: string
      refund_idempotency_key:
        type: string
      refundtxhash:
        type: string
    type: object
  api.refundResp:
    properties:
      message:
        type: string
      order_id:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Crypto payment processor API.
  title: OSPay API
  version: "1.0"
paths:
...
  /events/payment-detected:
    post:
      consumes:
      - application/json
      description: Notify the system of an on-chain payment for an order
      parameters:
      - description: Payment info
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/api.paymentDetectedReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.paymentDetectedResp'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.paymentDetectedResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Detect payment event
      tags:
      - events
  /merchants:
    post:
      consumes:
      - application/json
      description: Creates a new merchant and returns the merchant ID and API key
      parameters:
      - description: Merchant info
        in: body
        name: merchant
        required: true
        schema:
          $ref: '#/definitions/api.MerchantCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.MerchantCreateResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new merchant
      tags:
      - merchants
  /orders:
    post:
      consumes:
      - application/json
      - application/json
      description: |-
        Creates a new payment order for a merchant
        Creates a new payment order for a merchant
      parameters:
      - description: Order info
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/api.orderCreateReq'
      - description: Order info
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/api.orderCreateReq'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.orderCreateResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Create a new order
      tags:
      - orders
      - orders
  /orders/get:
    get:
      consumes:
      - application/json
      description: Returns order details for a given order ID
      parameters:
      - description: Order ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.orderGetResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get order by ID
      tags:
      - orders
  /orders/refund:
    post:
      consumes:
      - application/json
      description: Refunds a paid order by ID
      parameters:
      - description: Order ID
        in: query
        name: id
        required: true
        type: string
      - description: Refund info
        in: body
        name: refund
        required: true
        schema:
          $ref: '#/definitions/api.refundReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.refundResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Refund an order
      tags:
      - orders
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"